feat: Add ClientGesture class for gesture detection and handling
