+++ core/src/com/javaciv/client/ClientController.java
@@ -1,11 +1,3 @@
-/**
- * @file ClientController.java
- * @brief This file contains the ClientController class.
- * @author Théo Bessel
- * @date 20/04/2024
- * @version 1.0
- */
-
 package com.javaciv.client;
 
 import java.util.HashMap;
