+++ core/src/com/javaciv/client/ClientView.java
@@ -299,58 +299,7 @@ public class ClientView implements Screen {
         Gdx.input.setInputProcessor(this.inputMultiplexer);
 
         // Add a gesture detector to the input multiplexer
-        this.inputMultiplexer.addProcessor(new GestureDetector(new GestureDetector.GestureListener() {
-            @Override
-            public boolean tap(float x, float y, int count, int button) {
-                if (button == Input.Buttons.LEFT) {
-                    Leftclicked(x, y);
-                }
-                return false;
-            }
-
-            @Override
-            public boolean pan(float x, float y, float deltaX, float deltaY) {
-                if (Gdx.input.isButtonPressed(Input.Buttons.LEFT)) {
-                    camera.translate(-deltaX * camera.zoom, deltaY * camera.zoom);
-                    camera.update();
-                }
-                return false;
-            }
-
-            @Override
-            public boolean touchDown(float x, float y, int pointer, int button) {
-                return false;
-            }
-
-            @Override
-            public boolean longPress(float x, float y) {
-                return false;
-            }
-
-            @Override
-            public boolean fling(float velocityX, float velocityY, int button) {
-                return false;
-            }
-
-            @Override
-            public boolean panStop(float x, float y, int pointer, int button) {
-                return false;
-            }
-
-            @Override
-            public boolean zoom(float initialDistance, float distance) {
-                return false;
-            }
-
-            @Override
-            public boolean pinch(Vector2 initialPointer1, Vector2 initialPointer2, Vector2 pointer1, Vector2 pointer2) {
-                return false;
-            }
-
-            @Override
-            public void pinchStop() {
-            }
-        }));
+        this.inputMultiplexer.addProcessor(new GestureDetector(new ClientGesture(camera, controller, this)));
 
     }
